+++ src/screens/VisualTests/BuildResults.tsx
@@ -19,11 +19,11 @@ import {
   TestResult,
 } from "../../gql/graphql";
 import { LocalBuildProgress } from "../../types";
-import { Warnings } from "../../utils/Warnings";
 import { BuildEyebrow } from "./BuildEyebrow";
 import { FragmentStoryTestFields } from "./graphql";
 import { RenderSettings } from "./RenderSettings";
 import { SnapshotComparison } from "./SnapshotComparison";
+import { Warnings } from "./Warnings";
 
 interface BuildResultsProps {
   branch: string;
